+++ R/stats.R
@@ -258,3 +258,32 @@ site_stats <- function(x, site_name = 'XXX', year_range = NULL, filter_prop = 0.
   return(site.stats)
 }
 
+#' Percent scarred time series
+#'
+#' @param x An fhx object.
+#' @param injury_event Boolean indicating whether years with injury events should be considered as scars. Defaults to FALSE.
+#'
+#' @return A data.frame with four columns: \code{Year}, \code{NumRec} with the number of recording trees,
+#' \code{NumScars} with the number of fire scars and/or events, and \code{PercScarred} with the proportion of scars/events.
+#'
+#' @examples
+#' data("pgm")
+#' percent_scarred(pgm)
+#'
+#' @export
+percent_scarred <- function(x, injury_event=FALSE){
+  series_rec <- plyr::ddply(x, "series", burnr:::find_recording, injury_event=TRUE)
+  rec_count <- plyr::ddply(series_rec, "recording", count)
+
+  series_fs <- x[grepl('_fs', x$rec_type), ]
+  fs_count <- plyr::ddply(series_fs, "year", count)
+  if(injury_event) {
+    series_fs <- x[grepl('_fs', x$rec_type) | grepl('_fi', x$rec_type), ]
+    fs_count <- plyr::ddply(series_fs, "year", count)
+  }
+  out <- merge(rec_count, fs_count, by.x = 'recording', by.y = 'year', all=TRUE)
+  names(out) <- c('year', 'num_rec', 'num_scars')
+  out[is.na(out$num_scars), 'num_scars'] <- 0
+  out$percent_scarred <- round(out$num_scars / out$num_rec * 100, 0)
+  return(out)
+}