+++ tests/testthat/test-stats.R
@@ -94,8 +94,23 @@ test_that("series_mean_interval on multi-series fhx object", {
 })
 
 test_that("sample_depth on multi-series fhx object", {
-  sdepth <- subset(sample_depth(REF_MULTI), 
+  sdepth <- subset(sample_depth(REF_MULTI),
                    year %in% c(1366, 1436, 2011,2012))[['samp_depth']]
   expect_equal(sdepth, c(0, 1, 13, 2))
 })
 
+test_that("percent scarred works without injury", {
+  goal_1773_percent_scarred <- 50
+  victim <- percent_scarred(REF_MULTI)
+  expect_equal(victim[victim$year == 1773, ]$percent_scarred,
+               goal_1773_percent_scarred)
+  expect_equal(victim[victim$year == 1774, ]$percent_scarred,
+                0)
+})
+
+test_that("percent scarred works with injuries", {
+  goal_1806_percent_scarred <- 88
+  victim <- percent_scarred(REF_MULTI, injury_event = TRUE)
+  expect_equal(victim[victim$year == 1806, ]$percent_scarred,
+               goal_1806_percent_scarred)
+})